lots of tiny changes
started using intersection graph to find entities in blast radius instead of looping through every single dynamic rigid body and checking manually if it is in the blast like an idiot. switched to using shape casting instead of ray casting for the is_grounded check and ball deletion ray. fixed "pausing" the game so no input detection happens while the game is paused except for unpausing and fixed balls coming out of the bottom of the crosshair. also fixed printing how many balls there are only when you spawn them in so now when one is deleted it prints the count as well.  also learned you cant have multiple queries for a components transform but you can for their velocity which is weird